# recruitment-system/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recruitment System API",
        "description": "API documentation for the Recruitment System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/jobs/{job}/apply": {
            "post": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Apply for a job",
                "operationId": "5c9ebd5fcd5cd5976f428b41f92a99b6",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID of the job to apply for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "cv_file"
                                ],
                                "properties": {
                                    "cv_file": {
                                        "description": "CV file (pdf, doc, docx)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "cover_letter": {
                                        "description": "Cover letter",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/JobApplication"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad request (e.g., already applied, job closed)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-applications": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Get the current user's job applications",
                "operationId": "6c1bac5f237ec23c0d2616f7463cc6aa",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JobApplication"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications/{application}": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Get a specific job application",
                "operationId": "7a1d657a5dca6af231cbc690c6aa505c",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID of the job application",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobApplication"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications/{application}/status": {
            "patch": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Update the status of a job application",
                "operationId": "d43bbb31bfc11caaacf9d6b4d8cce402",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID of the job application",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "reviewed",
                                            "accepted",
                                            "rejected"
                                        ],
                                        "example": "reviewed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Application status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobApplication"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/job-poster/applications": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Get all applications for jobs posted by the current user",
                "operationId": "2d1d394a06c4dc7ddb2cce8387822713",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JobApplication"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{job}/applications": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Get all applications for a specific job",
                "operationId": "92b8b7f63987e7daf0a911ac619e15c7",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID of the job",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JobApplication"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications/{application}/cv": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Download the CV for a job application",
                "operationId": "aef377aa42effd3402d2ca1e3d16e01d",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID of the job application",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CV file download",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "CV not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "format": "email"
                    },
                    "phone_number": {
                        "description": "User phone number",
                        "type": "string"
                    },
                    "role": {
                        "description": "User role",
                        "type": "string",
                        "enum": [
                            "applicant",
                            "company",
                            "job_poster"
                        ]
                    },
                    "company_name": {
                        "description": "Company name",
                        "type": "string"
                    },
                    "company_type": {
                        "description": "Company type",
                        "type": "string",
                        "enum": [
                            "startup",
                            "national",
                            "multinational"
                        ]
                    },
                    "bio": {
                        "description": "User bio",
                        "type": "string"
                    },
                    "profile_picture_url": {
                        "description": "Profile picture URL",
                        "type": "string",
                        "format": "url"
                    },
                    "background_image_url": {
                        "description": "Background image URL",
                        "type": "string",
                        "format": "url"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "JobApplication": {
                "title": "Job Application",
                "description": "Job Application model",
                "properties": {
                    "id": {
                        "description": "Application ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "job_id": {
                        "description": "ID of the job applied for",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID of the applicant",
                        "type": "integer"
                    },
                    "cv_path": {
                        "description": "Path to the CV file",
                        "type": "string"
                    },
                    "cover_letter": {
                        "description": "Cover letter content",
                        "type": "string"
                    },
                    "status": {
                        "description": "Application status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "reviewed",
                            "accepted",
                            "rejected"
                        ]
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "job": {
                        "$ref": "#/components/schemas/Job"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Job": {
                "title": "Job",
                "description": "Job model",
                "properties": {
                    "id": {
                        "description": "Job ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Job title",
                        "type": "string"
                    },
                    "description": {
                        "description": "Job description",
                        "type": "string"
                    },
                    "requirements": {
                        "description": "Job requirements",
                        "type": "string"
                    },
                    "company_name": {
                        "description": "Company name",
                        "type": "string"
                    },
                    "min_salary": {
                        "description": "Minimum salary",
                        "type": "integer"
                    },
                    "max_salary": {
                        "description": "Maximum salary",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Job status",
                        "type": "string",
                        "enum": [
                            "active",
                            "closed"
                        ]
                    },
                    "user_id": {
                        "description": "ID of the user who posted the job",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}